Smarter Mushroom Spawning + Game Over Criteria
Added better positioning for mushroom spawning and a better way to
handle despawning them off the map as well. Also added the key feature
of singed running into the mushrooms and made the game exit when his
health was 0.

-- Still need to add a saftey precausion to mushrooms spawning on top of
mushrooms , similar to mushrooms onto temmos, lots of looping and
checking :/